% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbb_fns.R
\name{champ_season}
\alias{champ_season}
\title{Get NCAA championship winning seasons from a team's historical match data}
\usage{
champ_season(df, type)
}
\arguments{
\item{df}{Input data frame containing a team's historical match data.}

\item{type}{Match type}
}
\value{
Team's NCAA championship winning seasons (if any).
}
\description{
Get NCAA championship winning seasons from a team's historical match data
}
\examples{
champ_season(unc, "NCAA")
}

