% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbb_fns.R
\name{mbb_season_result}
\alias{mbb_season_result}
\title{Season match summary}
\usage{
mbb_season_result(df)
}
\arguments{
\item{df}{Input data frame containing a team's historical match data.}
}
\value{
A data frame containing number of games, wins, losses, win %, and loss %.
}
\description{
Season match summary
}
\examples{
mbb_season_result(unc)
}

